backend/requirements.txt

fastapi>=0.112
uvicorn[standard]>=0.30

qdrant-client>=1.9

sentence-transformers>=3.0
torch>=2.3
scikit-learn>=1.3

whoosh>=2.7

pymupdf>=1.24
pdfplumber>=0.11
python-docx>=1.1
pandas>=2.2
openpyxl>=3.1

python-multipart>=0.0.9
requests>=2.32
